fix: fixed average endpoint
